tests/wpt/tests/webmessaging/with-ports/018.html

<!doctype html>
<title>origin of the script that invoked the method, javascript:</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe src="javascript:''"></iframe>
<div id=log></div>
<script>
async_test(function() {
  window[0].postMessage('', '*', []);
  window[0].onmessage = this.step_func(function(e) {
    assert_equals(e.origin, location.protocol + '//' + location.host);
    assert_array_equals(e.ports, []);
    this.done();
  });
});
</script>
